doc: move sync server before provisioning
because the data sync service is delivered as a library used to write applications and the framework to run that library in OpenShift, putting the provisioning first causes the docs to describe an application you haven't year written.  By putting the server docs first developers can understand the context of the container the provisioning guide references.